requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.6.0
dlib==19.17.0
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.3
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
opencv-contrib-python==4.1.0.25
opencv-contrib-python-headless==4.1.0.25
opencv-python==4.1.0.25
opencv-python-headless==4.1.0.25
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyrsistent==0.15.2
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-dotenv==0.10.3
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.22.0
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
